requirements.txt
numpy~=1.21.6
aircv~=1.4.6
Pillow~=9.1.0
opencv-python~=4.5.5.64
PIC~=0.1.0
comtypes~=1.1.11
